Get Phenophases for Taxon

This function gets a list of phenophases that are applicable for a provided taxonomic grouping, e.g. family, order. Note that since a higher taxononmic order will aggregate individual species not every phenophase returned through this function will be applicable for every species belonging to that taxonomic group.

npn_get_phenophases_for_taxon(family_ids = NULL, order_ids = NULL,
  class_ids = NULL, date = NULL, return_all = 0, ...)

Arguments

family_ids
List of taxonomic family ids to search for.

order_ids
List of taxonomic order ids to search for.

class_ids
List of taxonomic class ids to search for

date
Specify the date of interst. For this function to return anything, either this value must be set of return_all must be 1.

return_all
Takes either 0 or 1 as input and defaults to 0. For this function to return anything, either this value must be set to 1 or date must be set.

...
Curl options passed on to GET

Details

It's also important to note that phenopphase definitions can change for individual species over time, so there's a need to specify either a date of interest, or to explicitely statea that the function should return all phenophases that were ever applicable for any species belonging to the specified taxonomic group.

When called, this function requires of these three paramaters, exactly one of family_ids, order_ids or class_ids to be set.