Get Species

Returns a complete list of all species information of species represented in the NPN database.

Returns information about a species based on the NPN's unique ID for that species

Search for species by state

Search NPN species information using a number of different parameters, which can be used in conjunction with one another, including: - Species on which a particular group or groups are actually collecting data - What species were observed in a given date range - What species were observed at a particular station or stations

npn_species(...)

npn_species_id(ids, ...)

npn_species_state(state, kingdom = NULL, ...)

npn_species_search(
  network = NULL,
  start_date = NULL,
  end_date = NULL,
  station_id = NULL,
  ...
)

Arguments

...
Curl options passed on to GET

ids
List of species ids for which to retrieve information

state
A US postal state code to filter results

kingdom
Filters results by taxonomic kingdom. Takes either 'Animalia' or 'Plantae'

network
filter species based on a list of unique identifiers of NPN groups that are actually observing data on the species. Takes a list of IDs

start_date
filter species by date observed. This sets the start date of the date range and must be used in conjunction with end_date

end_date
filter species by date observed. This sets the end date of the date range and must be used in conjunction with start_date

station_id
filter species by a list of unique site identifiers

Value

data.frame of species and their IDs

data.frame of the species' information

Examples

if (FALSE) {
npn_species()
npn_species_id(ids = 3)
}
if (FALSE) {
npn_species_state(state = "AZ")
npn_species_state(state = "AZ", kingdom = "Plantae")
}