Get Species

Returns a complete list of all species information of species represented in the NPN database.

Returns information about a species based on the NPN's unique ID for that species

Search for species by state

Search NPN species information using a number of different parameters, which can be used in conjunction with one another, including:

Species on which a particular group or groups are actually collecting data

What species were observed in a given date range

What species were observed at a particular station or stations

npn_species(...)

npn_species_id(ids, ...)

npn_species_state(state, kingdom = c("Animalia", "Plantae"), ...)

npn_species_search(
  network = NULL,
  start_date = NULL,
  end_date = NULL,
  station_id = NULL,
  ...
)

Arguments

...

Currently unused.

ids

Integer vector of species ids for which to retrieve information.

state

A US postal state code to filter results.

kingdom

Character vector of taxonomic kingdoms to filter results by. Valid values include 'Animalia' and 'Plantae'.

network

filter species based on identifiers of NPN groups that are actually observing data on the species. Takes a single numeric ID.

start_date

filter species by date observed. This sets the start date of the date range and must be used in conjunction with end_date.

end_date

filter species by date observed. This sets the end date of the date range and must be used in conjunction with start_date.

station_id

filter species by a numeric vector of unique site identifiers.

Value

A tibble with information on species in the NPN database and their IDs.

A tibble with information on species in the NPN database and their IDs, filtered by the species ID parameter.

A tibble with information on species in the NPN database whose distribution includes a given state.

A tibble with information on species in the NPN database filtered by partner group, dates and station/site IDs.

Examples

if (FALSE) { # \dontrun{
npn_species()
} # }
if (FALSE) { # \dontrun{
npn_species_id(ids = 3)
} # }
if (FALSE) { # \dontrun{
npn_species_state(state = "AZ")
# only return plants
npn_species_state(state = "AZ", kingdom = "Plantae")
} # }
if (FALSE) { # \dontrun{
species <- npn_species_search(
  start_date = "2013-01-01",
  end_date = "2013-05-15"
)
} # }